refactor: 달력 노출 영역 component 가독성 위한 정리
