+++ client/components/calendar/CalendarWrap.tsx
@@ -2,18 +2,16 @@ import styled from 'styled-components';
 
 import {useRecoilValue} from 'recoil';
 import {
-    targetStateState,
-    todayState,
     viewState
 } from '../../recoil/atoms';
 
 import {ViewType} from '../../utils/constants';
 
 import {DaysComponent} from './Days';
-import {DatesComponent} from './Dates';
 import {YearComponents} from './Year';
 import {WeekWrapComponent} from './WeekWrap';
 import {MonthWrapComponent} from './MonthWrap';
+import {TimelineComponent} from './Timeline';
 
 interface DaysType {
     type: string | null;
@@ -25,15 +23,18 @@ export const CalendarComponent = () => {
 
     return (
         <>
-            {(type !== ViewType.Year) && <StyledDaysWrap type={type}>
-                <DaysComponent>
-                    {(type === ViewType.Week || type === ViewType.Three) && <WeekWrapComponent type={type}/>}
-                </DaysComponent>
-            </StyledDaysWrap>}
+            {(type !== ViewType.Year) && <>
+                <StyledDaysWrap type={type}>
+                    <DaysComponent>
+                        {(type === ViewType.Week || type === ViewType.Three) && <WeekWrapComponent type={type}/>}
+                    </DaysComponent>
+                </StyledDaysWrap>
 
-            <DatesComponent>
-                {type === ViewType.Month && <MonthWrapComponent/>}
-            </DatesComponent>
+                <StyledDatesWrap type={type}>
+                    {type !== ViewType.Month && <TimelineComponent/>}
+                    {type === ViewType.Month && <MonthWrapComponent/>}
+                </StyledDatesWrap>
+            </>}
 
             {(type === ViewType.Year) && <YearComponents/>}
         </>
@@ -47,6 +48,20 @@ const StyledDaysWrap = styled.div <DaysType>`
     padding-left: 150px;
     box-sizing: border-box;
   `}
+  ul {
+    grid-template-columns: repeat(${props => props.type === ViewType.Three ? 3 : 7}, 1fr);
+  }
+`;
+
+const StyledDatesWrap = styled.div<{ type: string | null }>`
+  flex: 1;
+  display: grid;
+  grid-template-columns: ${props => props.type !== ViewType.Month ? '150px auto' : '1fr'};
+  width: 100%;
+  height: 100%;
+  overflow-x: hidden;
+  overflow-y: auto;
+  
   ul {
     grid-template-columns: repeat(${props => props.type === ViewType.Three ? 3 : 7}, 1fr);
   }
